+++ changelog.md
@@ -1,3 +1,10 @@
+**0.2.0-RC7**
+
+* New API /waves/payment returns senderPublicKey
+* New API /waves/create-signed-payment
+* /waves/external-payment deprecated. 
+  Use new /waves/broadcast-signed-payment.
+  
 **0.2.0-RC5**
 
 * /waves/external-payment returns error for incorrect recipient